requirements.txt
bs4
requests
tabula-py
pandas
numpy
lxml
aerosandbox
matplotlib
scipy
tqdm
